pyproject.toml
[project]
name = "embient"
version = "0.1.0"
description = "Embient CLI - AI-powered trading assistant"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11,<4.0"
dependencies = [
  "deepagents==0.3.8",
  "langchain>=1.2.3,<2.0.0",
  "langchain-openai>=1.1.7,<2.0.0",
  "langgraph-checkpoint-sqlite>=3.0.0,<4.0.0",
  "requests",
  "rich>=13.0.0",
  "prompt-toolkit>=3.0.52",
  "tavily-python",
  "python-dotenv",
  "daytona>=0.113.0",
  "modal>=0.65.0",
  "markdownify>=0.13.0",
  "runloop-api-client>=0.69.0",
  "pillow>=10.0.0",
  "pyyaml>=6.0",
  "textual>=1.0.0",
  "textual-autocomplete>=3.0.0",
  "aiosqlite>=0.19.0",
  "httpx>=0.28.0",
  "tenacity>=8.2.0",
]

[project.scripts]
embient = "embient:cli_main"

[dependency-groups]
test = [
    "pytest>=8.3.4",
    "pytest-asyncio>=0.25.3",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "pytest-socket>=0.7.0",
    "pytest-timeout>=2.3.1",
    "pytest-watcher>=0.3.4,<1.0.0",
    "pytest-xdist>=3.8.0,<4.0.0",
    "responses>=0.25.0",
    "build",
    "twine",
    "langchain-openai",
    "ruff>=0.9.7",
    "ty>=0.0.1,<1.0.0",
]

[build-system]
requires = ["setuptools>=73.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["embient*"]

[tool.setuptools.package-data]
embient = ["default_agent_prompt.md", "*.tcss", "py.typed"]

[tool.ruff]
line-length = 100
exclude = []

[tool.ruff.format]
docstring-code-format = true  # Formats code blocks in docstrings

[tool.ruff.lint]
select = [
    "ALL"  # Enable all rules by default
]
ignore = [
    "COM812",  # Messes with the formatter
    "ISC001",  # Messes with the formatter
    "PERF203", # Rarely useful
    "SLF001",  # Private member access
    "PLC0415", # Imports should be at the top. Not always desirable
    "PLR0913", # Too many arguments in function definition
    "PLC0414", # Inconsistent with how type checkers expect to be notified of intentional re-exports
    "C901",    # Too complex
]
unfixable = ["B028"]  # Rules that shouldn't be auto-fixed

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.ruff.lint.flake8-annotations]
allow-star-arg-any = true

[tool.ruff.lint.isort]
force-single-line = false
combine-as-imports = true
known-first-party = ["embient"]

[tool.ruff.lint.pydocstyle]
convention = "google"  # Google-style docstrings
ignore-var-parameters = true

[tool.ruff.lint.per-file-ignores]
"tests/*" = [
    "D1",      # Skip documentation rules in tests
    "S101",    # Allow asserts in tests
    "S311",    # Allow pseudo-random generators in tests
    "ANN201",  # Missing return type annotation
    "INP001",  # Implicit namespace package
    "PLR2004", # Magic value comparisons are fine in tests
]
"embient/agent.py" = ["E501", "PLR0912", "PLR2004", "SIM108"]
"embient/app.py" = ["E501", "PLR0915", "SIM108", "TC001", "TC002", "TC003"]
"embient/cli.py" = ["E501", "EM101", "EM102", "PLR0912", "PLR0915", "T201", "TC002", "TC003", "TRY003"]
"embient/clipboard.py" = ["BLE001", "PTH123", "S112", "TRY300"]
"embient/config.py" = ["D107", "E402", "E501", "ERA001", "FBT001", "FBT002", "RET503"]
"embient/file_ops.py" = ["ANN401", "BLE001", "D102", "PLR0912", "PLR0915"]
"embient/image_utils.py" = ["BLE001", "PERF401", "PLR0911", "PTH108", "PTH110", "PTH123", "PTH202", "S110", "S603", "S607", "SIM105"]
"embient/input.py" = ["ANN001", "ANN201", "BLE001", "D107", "FBT001", "FBT002", "PLR0912", "PLR0915", "TC002"]
"embient/integrations/daytona.py" = ["FIX002", "TD002", "TD003"]
"embient/integrations/modal_integration.py" = ["BLE001", "EM101", "EM102", "PLR0912", "S602", "S603", "S604", "S607", "TC002", "TC003", "TRY003"]
"embient/integrations/runloop.py" = ["B904", "F401"]
"embient/integrations/sandbox_factory.py" = ["BLE001", "EM101", "EM102", "S110", "TC003", "TRY003"]
"embient/core/agents_api.py" = ["EM101", "EM102", "TC003", "TRY003"]
"embient/core/api_credentials.py" = ["EM101", "EM102", "S603", "S607", "TC002", "TC003", "TRY003"]
"embient/core/env_manager.py" = ["EM101", "EM102", "TC002", "TC003", "TRY003"]
"embient/core/sessions.py" = ["ANN401", "BLE001", "EM101", "EM102", "FBT001", "PLR0913", "TC002", "TC003", "TRY003", "UP006"]
"embient/core/ui_display_types.py" = ["A003", "FBT001"]
"embient/local_context.py" = ["ARG002", "BLE001", "E501", "EM101", "EM102", "PERF401", "PLR0911", "PLR0912", "PLR0915", "S603", "S607", "TC002", "TC003", "TRY300"]
"embient/project_utils.py" = ["ERA001"]
"embient/shell.py" = ["PERF401", "S602"]
"embient/skills/commands.py" = ["ANN401", "E501", "FBT001", "FBT002", "TRY300"]
"embient/skills/load.py" = ["TC003"]
"embient/textual_adapter.py" = ["E501"]
"embient/tools.py" = ["ANN201", "BLE001", "E501", "E722", "FBT001", "FBT002", "PLR2004", "S113"]
"embient/ui.py" = ["ANN401", "ARG002", "BLE001", "D102", "EM102", "FBT001", "PLR0911", "PLR0912", "PLR0915", "S110", "TC002", "TC003", "TRY003", "TRY300"]
"embient/textual_app/components/messages.py" = ["TC002"]
"embient/widgets/approval.py" = ["ARG002", "B007", "D107", "RUF001", "TC002", "TC003"]
"embient/widgets/diff.py" = ["PLR0912"]
"embient/widgets/messages.py" = ["TC002"]
"embient/widgets/status.py" = ["PLR2004"]
"embient/widgets/tool_renderers.py" = ["D102", "PLR2004"]
"embient/main.py" = ["T201"]
"embient/skills.py" = ["ANN401", "D102", "D107", "EM101", "EM102", "F821", "PLR0913", "S102", "TC003", "TRY003"]
"embient/textual_app/app.py" = ["A001", "A002", "ANN202", "ARG002", "BLE001", "D102", "E501", "EM101", "EM102", "ERA001", "FBT001", "PLR0912", "PLR0913", "PLR0915", "S311", "SIM102", "T201", "TC002", "TC003", "TRY003"]
"embient/textual_app/conversation_service.py" = ["ANN401", "EM101", "EM102", "PLR0913", "RUF012", "TC003", "TRY003"]
"embient/textual_app/history_tab.py" = ["A002", "ANN202", "BLE001", "EM101", "EM102", "PLR0912", "RUF012", "TC002", "TC003", "TRY003"]
"embient/textual_app/input_box.py" = ["ANN202", "BLE001", "EM102", "PLR0912", "PLR0915", "TC003", "TRY003"]
"embient/textual_app/markdown_with_code_actions.py" = ["ANN202", "BLE001", "D102", "E501", "EM102", "SLF001", "TC002", "TC003", "TRY003"]
"embient/textual_app/running_shell_widget.py" = ["ANN202", "D102", "TC003"]
"embient/utils.py" = ["ANN401", "EM101", "EM102", "PLR0912", "S604", "S607", "TC002", "TC003", "TRY003"]
"tests/integration_tests/test_cli.py" = ["ANN001", "E501", "PLR0915", "S603", "S607"]
"tests/integration_tests/test_core_agents_api.py" = ["ANN001", "PLR0915"]
"tests/integration_tests/test_memory.py" = ["BLE001"]
"tests/integration_tests/test_runloop.py" = ["S106"]
"tests/integration_tests/test_sandbox_factory.py" = ["S108"]
"tests/integration_tests/test_sandbox_operations.py" = ["E501", "RUF001", "S108"]
"tests/integration_tests/test_ui_memory.py" = ["SIM118"]
"tests/unit_tests/skills/test_commands.py" = ["S108"]
"tests/unit_tests/test_api_credentials.py" = ["ANN001", "S106"]
"tests/unit_tests/test_autocomplete.py" = ["ANN001", "ARG002", "RET504"]
"tests/unit_tests/test_cli.py" = ["ANN001", "E501", "PLR0915"]
"tests/unit_tests/test_core_agents_api.py" = ["ANN001", "PLR0915"]
"tests/unit_tests/test_end_to_end.py" = ["ARG002", "E501", "RUF059"]
"tests/unit_tests/test_env_manager.py" = ["ANN001", "ANN202"]
"tests/unit_tests/test_local_context.py" = ["ANN001"]
"tests/unit_tests/test_sessions.py" = ["ANN001", "ANN202", "E501"]
"tests/unit_tests/test_version.py" = ["S603"]
"examples/skills/arxiv-search/arxiv_search.py" = ["BLE001", "D103", "TRY300"]
"examples/skills/skill-creator/scripts/init_skill.py" = ["ANN001", "ANN201", "BLE001", "D103", "D415", "PLR2004", "T201"]
"examples/skills/skill-creator/scripts/quick_validate.py" = ["ANN001", "ANN201", "D415", "E501", "N806", "PLR0911", "PLR0912", "PLR2004", "T201"]
"scripts/check_imports.py" = ["INP001", "BLE001"]

[tool.pytest.ini_options]
timeout = 10  # Default timeout for all tests (can be overridden per-test)

[tool.uv.sources]
deepagents = { path = "../deepagents/libs/deepagents", editable = true }
